% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preqc.R
\name{calc_preqc_metrics}
\alias{calc_preqc_metrics}
\title{Calculate preliminary quality control metrics}
\usage{
calc_preqc_metrics(empty_drops.out, fdr.threshold = 0.001)
}
\arguments{
\item{empty_drops.out}{The object returned by the \code{DropletUtils::emptyDrops()} function
that has been run on a raw (unfiltered) counts matrix.}

\item{fdr.threshold}{An FDR threshold for which to colour putative cell-containing
droplets versus putative empty droplets.}
}
\value{
A list with the following metrics:
\itemize{
\item knee : the calculated knee point by \code{DropletUtils::barcodeRanks()}
\item inflection : the calculated inflection point by \code{DropletUtils::barcodeRanks()}
\item n.gt_knee : number of droplets with number of UMIs greater than the knee point
\item n.putative_cell : number of droplets that are putative cells based on
specified FDR threshold
}
}
\description{
Calculates some metrics that are nice to have to assess quality of an
scRNA-seq sample, based on the output of the \code{DropletUtils::emptyDrops()}
function.
}
\examples{
suppressMessages(library(SingleCellExperiment))
sce.path <- system.file('extdata', 'sce', 'raw.rds', package = 'AssembleSC')
sce <- readRDS(sce.path)

empty_drops.out <- DropletUtils::emptyDrops(counts(sce), lower = 500)

calc_preqc_metrics(empty_drops.out, fdr.threshold = 0.001)

}
