% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preqc.R
\name{draw_kneeplot}
\alias{draw_kneeplot}
\title{Draw a kneeplot}
\usage{
draw_kneeplot(empty_drops.out, fdr.threshold = 0.001)
}
\arguments{
\item{empty_drops.out}{The object returned by the \code{DropletUtils::emptyDrops()} function
that has been run on a raw (unfiltered) counts matrix.}

\item{fdr.threshold}{An FDR threshold for which to colour putative cell-containing
droplets versus putative empty droplets.}
}
\value{
A ggplot2 object with plenty of customization.
}
\description{
Creates the classic kneeplot.
}
\examples{
suppressMessages(library(SingleCellExperiment))
sce.path <- system.file('extdata', 'sce', 'raw.rds', package = 'AssembleSC')
sce <- readRDS(sce.path)

empty_drops.out <- DropletUtils::emptyDrops(counts(sce), lower = 500)

draw_kneeplot(empty_drops.out, fdr.threshold = 0.001)

}
\seealso{
\code{\link[=calc_preqc_metrics]{calc_preqc_metrics()}} for key calculated preliminary metrics used in this plot.
}
