% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{ingest_kallisto}
\alias{ingest_kallisto}
\title{Ingest kallisto output}
\usage{
ingest_kallisto(
  dir.all,
  .id = "auto",
  strategy = "full_length",
  tx2gene = NULL
)
}
\arguments{
\item{dir.all}{A string with the path pointing to the directory \emph{above} the individual sample directories,
the one containing \emph{all} samples (cells) from the run; this is due to between-sample
normalization that needs to be performed prior to analysis. Required.}

\item{.id}{A character vector specifying the sample ids. Default is 'auto', which deduces the
sample id based on the folder names within the directory specified by \code{dir.all}.}

\item{strategy}{A string specifying the sequencing protocol's strategy, used to instruct
the normalization approach. Default is '3prime_tagged'.}

\item{tx2gene}{A two-column data.frame linking transcript id (column 1) to gene id (column 2).
Required for gene-level summarisation, when \code{txOut = FALSE}.}
}
\description{
Shortcircuits the regular \code{ingest()} function to return a \code{SingleCellExperiment}
object derived from regular kallisto output. This is used typically for processing
plate-based data, where each well corresponds to a single cell, and each cell
is in its own directory, e.g. sample folder.
}
\details{
This function utilizes under the hood the \code{tximport::tximport()} function to
read in the counts data, performing scaling, gene summarisation, and the like
as specified. For more details, see \code{?tximport::tximport()}.

Note that depending on the sequencing protocol, different counts-normalization
strategies may need to be used. For example, for full-length 5'->3' sequencing,
gene/transcript length normalization is a common step, as opposed to 3' targeted
strategy which does not require this step.

For 3' tagged RNA-seq, correcting the counts for gene length will induce a bias
in your analysis, because the counts do not have a length bias. Thus, it is
recommended to use the original counts, as a counts matrix.
For full-length sequencing, its recommended to set \code{countsFromAbundance} to

For more information, see:
\href{https://bioconductor.org/packages/devel/bioc/vignettes/tximport/inst/doc/tximport.html#downstream_dge_in_bioconductor}{Downstream DGE in Bioconductor from tximport} and \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6178912/}{Swimming downstream}.
}
\examples{
\dontrun{
dir <- system.file('extdata', 'aligners', 'kallisto', 'Sample_A1', package = 'AssembleSC')
t2g.file <- system.file('extdata', 'metadata', 'tx2gene_human.tsv', package = 'AssembleSC')
t2g <- readr::read_tsv(t2g.file, col_names = FALSE)

ingest_kallisto(dir)
}

}
\seealso{
Other ingest: 
\code{\link{ingest}()}
}
\concept{ingest}
